change project structure: separate headers in include/ and sources in src/; add some dummies in new dirs
