+++ gui-main.cpp
@@ -1,0 +1,5 @@
+#include "gui/GUIDummy.h"
+
+int main() {
+    return 0;
+}
