+++ include/cli/ConsoleMessage.h
@@ -1,0 +1,11 @@
+#pragma once
+
+/*
+*   Example of a console header (dummmy, may be deleted)
+*/
+
+#include <iostream>
+
+namespace erconv {
+    void HelloConsoleMessage();
+}
