+++ include/gui/GUIDummy.h
@@ -1,0 +1,5 @@
+#pragma once
+
+/*
+*   GUIDummy (may be deleted)
+*/
