+++ src/base/Entity.cpp
@@ -1,4 +1,4 @@
-#include "Entity.h"
+#include "base/Entity.h"
 
 /*  
     Simple method to added new field in Entity
