+++ src/cli/ConsoleMessage.cpp
@@ -1,0 +1,9 @@
+#include "cli/ConsoleMessage.h"
+
+/*
+*   Example of a console cpp (dummy, may be deleted)
+*/
+
+void erconv::HelloConsoleMessage() {
+    std::cout << "Hello, Console!" << std::endl;
+}
