+++ src/gui/GUIDummy.cpp
@@ -1,0 +1,5 @@
+#include "gui/GUIDummy.h"
+
+/*
+*   GUIDummy (may be deleted)
+*/
