+++ tests/er-model-test.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include "Model.h"
+#include "base/Model.h"
 
 namespace erconv {
